Completed SneakerService and WhiskeyService plus tests

+++ src/main/java/models/Sneaker.java
@@ -5,13 +5,13 @@ public class Sneaker {
     private String name;
     private String brand;
     private String sport;
-    private int size;
+    private Double size;
     private int qty;
     private float price;
 
     public Sneaker() {}
 
-    public Sneaker(int id, String name, String brand, String sport, int size, int qty, float price) {
+    public Sneaker(int id, String name, String brand, String sport, Double size, int qty, float price) {
         this.id = id;
         this.name = name;
         this.brand = brand;
@@ -54,11 +54,11 @@ public class Sneaker {
         return sport;
     }
 
-    public void setSize(int size) {
+    public void setSize(Double size) {
         this.size = size;
     }
 
-    public int getSize() {
+    public Double getSize() {
         return size;
     }
